requirements.txt
flask
celery
redis
mysql-connector-python
prometheus-client
pytest
pytest-cov
requests
PyYAML
pre-commit
black
isort
flake8
Pillow
